Put bpf2bpf caller in a named section
Libbpf logic is that if there are multiple programs then any
programs in the .text section are subprograms to be skipped
when enumerating programs.

So update the sample to follow the convention that the non-subprogram
is in a section other than ".text".  This also enables testing the
verifier's ability to load programs that call subprograms in another
section.

Signed-off-by: Dave Thaler <dthaler1968@gmail.com>